requirements.txt
Django==5.0
djangorestframework==3.15
psycopg2-binary==2.9.1
django-allauth==0.44.0
dj-rest-auth==6.0.0
stripe==2.60.0
requests==2.25.1
django-environ==0.11.2
